+++ src/image.zig
@@ -6,7 +6,7 @@ const Interval = @import("interval.zig").Interval;
 const Vec3 = vec3.Vec3;
 const Color = vec3.Color;
 
-pub fn linear_to_gamma(x: f32) f32 {
+pub fn linear_to_gamma(x: f64) f64 {
     return @sqrt(x);
 }
 
@@ -63,7 +63,7 @@ pub const PPMImage = struct {
         var g = c.y;
         var b = c.z;
 
-        const scale = 1.0 / @as(f32, @floatFromInt(samples_per_pixel));
+        const scale = 1.0 / @as(f64, @floatFromInt(samples_per_pixel));
         r *= scale;
         g *= scale;
         b *= scale;
